PROG-37: update mock data
